requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
Pillow==10.1.0
psycopg==3.2.5
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uuid==1.30
